+++ src/libraries/Common/src/Interop/Unix/Interop.PathStackBufferSize.cs
@@ -5,5 +5,5 @@ internal static partial class Interop
 {
     // Unix max paths are typically 1K or 4K UTF-8 bytes, 256 should handle the majority of paths
     // without putting too much pressure on the stack.
-    internal const int PathStackBufferSize = 256;
+    internal const int DefaultPathBufferSize = 256;
 }
