+++ src/libraries/Common/src/Interop/Unix/System.Native/Interop.ReadLink.cs
@@ -4,6 +4,7 @@
 using System.Runtime.InteropServices;
 using System.Buffers;
 using System.Text;
+using System;
 
 internal static partial class Interop
 {
@@ -20,42 +21,52 @@ internal static partial class Interop
         /// Returns the number of bytes placed into the buffer on success; bufferSize if the buffer is too small; and -1 on error.
         /// </returns>
         [DllImport(Libraries.SystemNative, EntryPoint = "SystemNative_ReadLink", SetLastError = true)]
-        private static extern int ReadLink(string path, byte[] buffer, int bufferSize);
+        private static extern int ReadLink(ref byte path, byte[] buffer, int bufferSize);
 
         /// <summary>
         /// Takes a path to a symbolic link and returns the link target path.
         /// </summary>
-        /// <param name="path">The path to the symlink</param>
-        /// <returns>
-        /// Returns the link to the target path on success; and null otherwise.
-        /// </returns>
-        public static string? ReadLink(string path)
+        /// <param name="path">The path to the symlink.</param>
+        /// <returns>Returns the link to the target path on success; and null otherwise.</returns>
+        internal static string? ReadLink(ReadOnlySpan<char> path)
         {
-            int bufferSize = PathStackBufferSize;
-            while (true)
+            int bufferSize = Math.Min(path.Length, DefaultPathBufferSize);
+            var converter = new ValueUtf8Converter(stackalloc byte[bufferSize]);
+            try
             {
-                byte[] buffer = ArrayPool<byte>.Shared.Rent(bufferSize);
-                try
+                while (true)
                 {
-                    int resultLength = Interop.Sys.ReadLink(path, buffer, buffer.Length);
-                    if (resultLength < 0)
+                    byte[] buffer = ArrayPool<byte>.Shared.Rent(bufferSize);
+                    try
                     {
-                        // error
-                        return null;
-                    }
-                    else if (resultLength < buffer.Length)
-                    {
-                        // success
-                        return Encoding.UTF8.GetString(buffer, 0, resultLength);
-                    }
-                }
-                finally
-                {
-                    ArrayPool<byte>.Shared.Return(buffer);
-                }
+                        int resultLength = Interop.Sys.ReadLink(
+                            ref MemoryMarshal.GetReference(converter.ConvertAndTerminateString(path)),
+                            buffer,
+                            buffer.Length);
 
-                // buffer was too small, loop around again and try with a larger buffer.
-                bufferSize *= 2;
+                        if (resultLength < 0)
+                        {
+                            // error
+                            return null;
+                        }
+                        else if (resultLength < buffer.Length)
+                        {
+                            // success
+                            return Encoding.UTF8.GetString(buffer, 0, resultLength);
+                        }
+                    }
+                    finally
+                    {
+                        ArrayPool<byte>.Shared.Return(buffer);
+                    }
+
+                    // buffer was too small, loop around again and try with a larger buffer.
+                    bufferSize *= 2;
+                }
+            }
+            finally
+            {
+                converter.Dispose();
             }
         }
     }
